♻️ Rename setupProject to be consistent
