+++ src/commands/setup-project.ts
